Refactor: write-behind로 포인트 관리를 수행하는 코드는 캐시와 사용자 인터페이스 사이의 중재역할을 수행하고 있으므로 application 계층으로 이동
